requirements.txt
Django==3.0.5
django-cors-headers==2.4.0
django-debug-toolbar==1.9.1
django-redis==4.10.0
django-user-agents==0.3.2
django-webpack-loader==0.6.0
djangorestframework==3.8.2
dnspython==1.16.0
django_heroku == 0.3.1
requests==2.20.1
requests-oauthlib==1.3.0
PyCrypto==2.6.1
gunicorn==19.0.0
whitenoise == 5.0.1
faker == 4.1.1
python-dateutil == 2.8.1
text-unidecode==1.3
six == 1.12.0